move adjacency_matrix agg to its own package
