+++ modules/aggregations/src/main/java/module-info.java
@@ -12,6 +12,7 @@ module org.elasticsearch.aggs {
     requires org.elasticsearch.xcontent;
     requires org.apache.lucene.core;
 
-    exports org.elasticsearch.aggregations.bucket;
+    exports org.elasticsearch.aggregations.bucket.histogram;
+    exports org.elasticsearch.aggregations.bucket.adjacency;
 
 }
